Change layout. Restructure files.
