+++ Utility/EsperTemplateUtil.py
@@ -7,6 +7,7 @@ import wx
 import time
 
 import Utility.wxThread as wxThread
+import Utility.EventUtility as eventUtil
 import Common.Globals as Globals
 
 from Utility.ApiToolLogging import ApiToolLog
@@ -89,7 +90,7 @@ class EsperTemplateUtil:
             if templateExist:
                 templateFound["id"] = templateExist[0]["id"]
                 postEventToFrame(
-                    wxThread.myEVT_CONFIRM_CLONE_UPDATE,
+                    eventUtil.myEVT_CONFIRM_CLONE_UPDATE,
                     (
                         self,
                         self.toApi,
@@ -102,7 +103,7 @@ class EsperTemplateUtil:
             else:
                 templateFound["id"] = maxId + 1
                 postEventToFrame(
-                    wxThread.myEVT_CONFIRM_CLONE,
+                    eventUtil.myEVT_CONFIRM_CLONE,
                     (
                         self,
                         self.toApi,
@@ -114,7 +115,7 @@ class EsperTemplateUtil:
                 )
         else:
             postEventToFrame(
-                wxThread.myEVT_LOG, "Template was not found. Check arguements."
+                eventUtil.myEVT_LOG, "Template was not found. Check arguements."
             )
 
     @api_tool_decorator()
@@ -127,7 +128,7 @@ class EsperTemplateUtil:
             templateFound, toDeviceGroups, allDeviceGroupId
         )
         if not found:
-            postEventToFrame(wxThread.myEVT_LOG, "Creating new device group...")
+            postEventToFrame(eventUtil.myEVT_LOG, "Creating new device group...")
             res = createDeviceGroupForHost(
                 self.getEsperConfig(self.toApi, self.toKey),
                 self.toEntId,
@@ -143,7 +144,7 @@ class EsperTemplateUtil:
             else:
                 templateFound["template"]["device_group"] = allDeviceGroupId
                 postEventToFrame(
-                    wxThread.myEVT_LOG,
+                    eventUtil.myEVT_LOG,
                     "Failed to recreate Device Group, using All Device group!",
                 )
                 wx.MessageBox(
@@ -155,7 +156,9 @@ class EsperTemplateUtil:
     @api_tool_decorator()
     def processWallpapers(self, templateFound):
         if self.toApi and self.toKey and self.toEntId:
-            postEventToFrame(wxThread.myEVT_LOG, "Processing wallpapers in template...")
+            postEventToFrame(
+                eventUtil.myEVT_LOG, "Processing wallpapers in template..."
+            )
             if templateFound["template"]["brand"]:
                 bgList = []
                 for bg in templateFound["template"]["brand"]["wallpapers"]:
@@ -203,16 +206,16 @@ class EsperTemplateUtil:
                 )
                 files = {"image_file": open("wallpaper.jpeg", "rb")}
                 postEventToFrame(
-                    wxThread.myEVT_LOG, "Attempting to upload wallpaper..."
+                    eventUtil.myEVT_LOG, "Attempting to upload wallpaper..."
                 )
                 resp = performPostRequestWithRetry(
                     url, headers=headers, data=payload, files=files
                 )
                 if resp.ok:
-                    postEventToFrame(wxThread.myEVT_LOG, "Wallpaper upload Succeeded!")
+                    postEventToFrame(eventUtil.myEVT_LOG, "Wallpaper upload Succeeded!")
                     json_resp = resp.json()
                 else:
-                    postEventToFrame(wxThread.myEVT_LOG, "Wallpaper upload Failed!")
+                    postEventToFrame(eventUtil.myEVT_LOG, "Wallpaper upload Failed!")
                     wx.MessageBox(
                         "Wallpaper upload Failed! Source: %s" % bg["url"],
                         style=wx.OK | wx.ICON_ERROR,
@@ -392,7 +395,7 @@ class EsperTemplateUtil:
                             }
                         )
                         postEventToFrame(
-                            wxThread.myEVT_LOG,
+                            eventUtil.myEVT_LOG,
                             "Added the '%s' app to the template"
                             % app["applicationName"],
                         )
@@ -401,7 +404,7 @@ class EsperTemplateUtil:
     def uploadMissingApk(self, app, template, newTemplate, config, entId):
         try:
             postEventToFrame(
-                wxThread.myEVT_LOG,
+                eventUtil.myEVT_LOG,
                 "Attempting to download %s to upload to endpoint" % app["packageName"],
             )
             file = "%s.apk" % app["applicationName"]
@@ -427,7 +430,7 @@ class EsperTemplateUtil:
             print(e)
             ApiToolLog().LogError(e)
             postEventToFrame(
-                wxThread.myEVT_LOG,
+                eventUtil.myEVT_LOG,
                 "To Enterprise is missing app, %s, not adding to template"
                 % app["applicationName"],
             )
@@ -534,7 +537,7 @@ class EsperTemplateUtil:
             if hasattr(resp, "status_code"):
                 if resp.status_code > 299:
                     postEventToFrame(
-                        wxThread.myEVT_MESSAGE_BOX,
+                        eventUtil.myEVT_MESSAGE_BOX,
                         (str(json_resp), wx.ICON_ERROR),
                     )
             return json_resp
