+++ Utility/GridActionUtility.py
@@ -1,0 +1,689 @@
+#!/usr/bin/env python
+
+import time
+
+import esperclient
+from Common import Globals
+
+from Utility.ApiToolLogging import ApiToolLog
+from Utility.Resource import (
+    joinThreadList,
+    limitActiveThreads,
+    postEventToFrame,
+    splitListIntoChunks,
+)
+from Utility import wxThread
+import Utility.EsperAPICalls as apiCalls
+
+from Common.decorator import api_tool_decorator
+from Common.enum import GridActions
+import Common.Globals as Globals
+import Utility.EventUtility as eventUtil
+
+
+@api_tool_decorator()
+def iterateThroughGridRows(frame, action):
+    """Iterates Through Each Device in the Displayed Grid And Performs A Specified Action"""
+    if action == GridActions.MODIFY_ALIAS_AND_TAGS.value:
+        modifyDevice(frame)
+    if (
+        action == GridActions.SET_APP_STATE_DISABLE.value
+        or action == GridActions.SET_APP_STATE_HIDE.value
+        or action == GridActions.SET_APP_STATE_SHOW.value
+    ):
+        setAppStateForAllAppsListed(action)
+    if action == 50:
+        setAppStateForSpecificAppListed(action)
+    if action == GridActions.MOVE_GROUP.value:
+        relocateDeviceToNewGroup(frame)
+    if action == GridActions.INSTALL_LATEST_APP.value:
+        pass
+    if action == GridActions.UNINSTALL_LISTED_APP.value:
+        pass
+
+
+@api_tool_decorator()
+def modifyDevice(frame):
+    """ Start a thread that will attempt to modify device data """
+    t = wxThread.GUIThread(
+        frame,
+        executeDeviceModification,
+        args=(frame),
+        eventType=None,
+        name="executeDeviceModification",
+    )
+    t.start()
+    return t
+
+
+@api_tool_decorator()
+def executeDeviceModification(frame, maxAttempt=Globals.MAX_RETRY):
+    """ Attempt to modify device data according to what has been changed in the Grid """
+    tagsFromGrid, rowTaglist = frame.gridPanel.getDeviceTagsFromGrid()
+    aliasDic = frame.gridPanel.getDeviceAliasFromList()
+    frame.gauge.SetValue(1)
+    maxGaugeAction = len(tagsFromGrid.keys()) + len(aliasDic.keys())
+
+    devices = []
+    for row in rowTaglist.keys():
+        entry = rowTaglist[row]
+
+        api_instance = esperclient.DeviceApi(
+            esperclient.ApiClient(Globals.configuration)
+        )
+        api_response = None
+        identifier = None
+        for attempt in range(maxAttempt):
+            try:
+                if "esperName" in entry.keys():
+                    identifier = entry["esperName"]
+                    api_response = api_instance.get_all_devices(
+                        Globals.enterprise_id,
+                        name=identifier,
+                        limit=Globals.limit,
+                        offset=Globals.offset,
+                    )
+                elif "sn" in entry.keys():
+                    identifier = entry["sn"]
+                    api_response = api_instance.get_all_devices(
+                        Globals.enterprise_id,
+                        serial=identifier,
+                        limit=Globals.limit,
+                        offset=Globals.offset,
+                    )
+                elif "csn" in entry.keys():
+                    identifier = entry["csn"]
+                    api_response = api_instance.get_all_devices(
+                        Globals.enterprise_id,
+                        search=identifier,
+                        limit=Globals.limit,
+                        offset=Globals.offset,
+                    )
+                ApiToolLog().LogApiRequestOccurrence(
+                    "getAllDevices",
+                    api_instance.get_all_devices,
+                    Globals.PRINT_API_LOGS,
+                )
+                break
+            except Exception as e:
+                if attempt == maxAttempt - 1:
+                    ApiToolLog().LogError(e)
+                    raise e
+                time.sleep(Globals.RETRY_SLEEP)
+        if api_response:
+            devices += api_response.results
+            api_response = None
+        else:
+            postEventToFrame(
+                eventUtil.myEVT_LOG,
+                "---> ERROR: Failed to find device with identifer: %s" % identifier,
+            )
+
+    splitResults = splitListIntoChunks(devices)
+
+    threads = []
+    for chunk in splitResults:
+        t = wxThread.GUIThread(
+            frame,
+            processDeviceModificationForList,
+            args=(frame, chunk, tagsFromGrid, aliasDic, maxGaugeAction),
+            name="processDeviceModificationForList",
+        )
+        threads.append(t)
+        t.start()
+        limitActiveThreads(threads)
+
+    t = wxThread.GUIThread(
+        frame,
+        wxThread.waitTillThreadsFinish,
+        args=(tuple(threads), -1, -1, 2),
+        name="waitTillThreadsFinish",
+    )
+    t.start()
+
+
+@api_tool_decorator()
+def processDeviceModificationForList(
+    frame, chunk, tagsFromGrid, aliasDic, maxGaugeAction
+):
+    changeSucceeded = 0
+    succeeded = 0
+    numNewName = 0
+    status = []
+    tagStatus = []
+    for device in chunk:
+        t = wxThread.GUIThread(
+            frame,
+            changeTagsForDevice,
+            args=(device, tagsFromGrid, frame, maxGaugeAction),
+            name="changeTagsForDevice",
+        )
+        t.start()
+        t2 = wxThread.GUIThread(
+            frame,
+            changeAliasForDevice,
+            args=(device, aliasDic, frame, maxGaugeAction),
+            name="changeAliasForDevice",
+        )
+        t2.start()
+        joinThreadList([t, t2])
+        if t.result:
+            changeSucceeded += t.result[0]
+            if len(t.result) > 1 and t.result[1]:
+                tagStatus.append(t.result[1])
+        if t2.result:
+            numNewName += t2.result[0]
+            succeeded += t2.result[1]
+            if len(t2.result) > 2 and t2.result[2]:
+                status.append(t2.result[2])
+
+    status += tagStatus
+
+    return (changeSucceeded, succeeded, numNewName, tagsFromGrid, status)
+
+
+@api_tool_decorator()
+def changeAliasForDevice(device, aliasDic, frame, maxGaugeAction):
+    numNewName = 0
+    succeeded = 0
+    logString = ""
+    status = None
+    # Alias modification
+    if (
+        device.device_name in aliasDic.keys()
+        or (
+            "serialNumber" in device.hardware_info
+            and device.hardware_info["serialNumber"] in aliasDic.keys()
+        )
+        or (
+            "customSerialNumber" in device.hardware_info
+            and device.hardware_info["customSerialNumber"] in aliasDic.keys()
+        )
+    ):
+        newName = None
+        if device.device_name in aliasDic:
+            newName = aliasDic[device.device_name]
+        elif (
+            "serialNumber" in device.hardware_info
+            and device.hardware_info["serialNumber"] in aliasDic
+        ):
+            newName = aliasDic[device.hardware_info["serialNumber"]]
+        elif (
+            "customSerialNumber" in device.hardware_info
+            and device.hardware_info["customSerialNumber"] in aliasDic
+        ):
+            newName = aliasDic[device.hardware_info["customSerialNumber"]]
+        logString = str(
+            "--->" + str(device.device_name) + " : " + str(newName) + "--->"
+        )
+        if not newName and not device.alias_name:
+            return
+        if newName != str(device.alias_name):
+            numNewName += 1
+            status = ""
+            try:
+                ignoreQueued = False if Globals.REACH_QUEUED_ONLY else True
+                status = apiCalls.setdevicename(frame, device.id, newName, ignoreQueued)
+            except Exception as e:
+                ApiToolLog().LogError(e)
+            if "Success" in str(status):
+                logString = logString + " <success>"
+                succeeded += 1
+            elif "Queued" in str(status):
+                logString = logString + " <Queued> Make sure device is online."
+                postEventToFrame(eventUtil.myEVT_ON_FAILED, (device, "Queued"))
+            elif "Scheduled" in str(status):
+                logString = logString + " <Scheduled> Make sure device is online."
+                postEventToFrame(eventUtil.myEVT_ON_FAILED, (device, "Scheduled"))
+            else:
+                logString = logString + " <failed>"
+                postEventToFrame(eventUtil.myEVT_ON_FAILED, device)
+        else:
+            logString = logString + " (Alias Name already set)"
+        if "Success" in logString or "Queued" in logString:
+            postEventToFrame(eventUtil.myEVT_UPDATE_GRID_CONTENT, (device, "alias"))
+        postEventToFrame(
+            eventUtil.myEVT_UPDATE_GAUGE,
+            int(frame.gauge.GetValue() + 1 / maxGaugeAction * 100),
+        )
+        postEventToFrame(eventUtil.myEVT_LOG, logString)
+    return (numNewName, succeeded, status)
+
+
+@api_tool_decorator()
+def changeTagsForDevice(device, tagsFromGrid, frame, maxGaugeAction):
+    # Tag modification
+    changeSucceeded = 0
+    if (
+        device.device_name in tagsFromGrid.keys()
+        or (
+            "serialNumber" in device.hardware_info
+            and device.hardware_info["serialNumber"] in tagsFromGrid.keys()
+        )
+        or (
+            "customSerialNumber" in device.hardware_info
+            and device.hardware_info["customSerialNumber"] in tagsFromGrid.keys()
+        )
+    ):
+        tagsFromCell = None
+        key = None
+        if device.device_name in tagsFromGrid:
+            key = device.device_name
+        elif (
+            "serialNumber" in device.hardware_info
+            and device.hardware_info["serialNumber"] in tagsFromGrid
+        ):
+            key = device.hardware_info["serialNumber"]
+        elif (
+            "customSerialNumber" in device.hardware_info
+            and device.hardware_info["customSerialNumber"] in tagsFromGrid
+        ):
+            key = device.hardware_info["customSerialNumber"]
+        tagsFromCell = tagsFromGrid[key]
+        try:
+            tags = apiCalls.setdevicetags(device.id, tagsFromCell)
+        except Exception as e:
+            ApiToolLog().LogError(e)
+        if tags == tagsFromGrid[key]:
+            changeSucceeded += 1
+        postEventToFrame(eventUtil.myEVT_UPDATE_GRID_CONTENT, (device, "tags"))
+        postEventToFrame(eventUtil.myEVT_UPDATE_TAG_CELL, (device.device_name, tags))
+        postEventToFrame(
+            eventUtil.myEVT_UPDATE_GAUGE,
+            int(frame.gauge.GetValue() + 1 / maxGaugeAction * 100),
+        )
+        status = {"Device Identifier": key, "Tags": tags}
+    return changeSucceeded, status
+
+
+@api_tool_decorator()
+def setAppStateForAllAppsListed(state, maxAttempt=Globals.MAX_RETRY):
+    deviceIdentifers = Globals.frame.gridPanel.getDeviceIdentifersFromGrid()
+    devices = getDevicesFromGrid(deviceIdentifers=deviceIdentifers)
+    if devices:
+        threads = []
+        for device in devices:
+            if (
+                device.device_name in deviceIdentifers
+                or (
+                    "serialNumber" in device.hardware_info
+                    and device.hardware_info["serialNumber"] in deviceIdentifers
+                )
+                or (
+                    "customSerialNumber" in device.hardware_info
+                    and device.hardware_info["customSerialNumber"] in deviceIdentifers
+                )
+            ):
+                t = wxThread.GUIThread(
+                    Globals.frame,
+                    setAllAppsState,
+                    args=(Globals.frame, device, state),
+                    name="setAllAppsState",
+                )
+                threads.append(t)
+                t.start()
+                limitActiveThreads(threads)
+        t = wxThread.GUIThread(
+            Globals.frame,
+            wxThread.waitTillThreadsFinish,
+            args=(tuple(threads), state, -1, 4),
+            name="waitTillThreadsFinish%s" % state,
+        )
+        t.start()
+
+
+@api_tool_decorator()
+def setAllAppsState(frame, device, state):
+    stateStatuses = []
+    # for device in chunk:
+    _, resp = apiCalls.getdeviceapps(device.id, False, Globals.USE_ENTERPRISE_APP)
+    for app in resp["results"]:
+        stateStatus = None
+        package_name = None
+        app_version = None
+        if "application" in app:
+            package_name = app["application"]["package_name"]
+            app_version = app["application"]["version"]["version_code"]
+            if app["application"]["package_name"] in Globals.BLACKLIST_PACKAGE_NAME:
+                continue
+        else:
+            package_name = app["package_name"]
+            app_version = app["version_code"]
+            if app["package_name"] in Globals.BLACKLIST_PACKAGE_NAME:
+                continue
+        if state == GridActions.SET_APP_STATE_DISABLE.value:
+            stateStatus = apiCalls.setAppState(
+                device.id,
+                package_name,
+                appVer=app_version,
+                state="DISABLE",
+            )
+        if state == GridActions.SET_APP_STATE_HIDE.value:
+            stateStatus = apiCalls.setAppState(
+                device.id,
+                package_name,
+                appVer=app_version,
+                state="HIDE",
+            )
+        if state == GridActions.SET_APP_STATE_SHOW.value:
+            stateStatus = apiCalls.setAppState(
+                device.id,
+                package_name,
+                appVer=app_version,
+                state="SHOW",
+            )
+        if stateStatus and hasattr(stateStatus, "state"):
+            entry = {
+                "Device Name": device.device_name,
+                "Device id": device.id,
+                "State Status": stateStatus.state,
+            }
+            if hasattr(stateStatus, "reason"):
+                entry["Reason"] = stateStatus.reason
+            stateStatuses.append(entry)
+        else:
+            stateStatuses.append(
+                {
+                    "Device Name": device.device_name,
+                    "Device id": device.id,
+                    "State Status": stateStatus,
+                }
+            )
+    return stateStatuses
+
+
+def setAppStateForSpecificAppListed(action, maxAttempt=Globals.MAX_RETRY):
+    api_response = getDevicesFromGrid()
+    state = None
+    if action == 50:
+        state == "HIDE"
+
+    # deviceIdentifers = Globals.frame.gridPanel.getDeviceIdentifersFromGrid()
+    appList = Globals.frame.gridPanel.getDeviceAppFromGrid()
+    if api_response:
+        tempRes = []
+        for device in api_response.results:
+            for deviceIds in appList.keys():
+                if (
+                    device.device_name in deviceIds
+                    or (
+                        "serialNumber" in device.hardware_info
+                        and device.hardware_info["serialNumber"] in deviceIds
+                    )
+                    or (
+                        "customSerialNumber" in device.hardware_info
+                        and device.hardware_info["customSerialNumber"] in deviceIds
+                    )
+                ):
+                    tempRes.append(device)
+        if tempRes:
+            api_response.results = tempRes
+        threads = []
+        for device in api_response.results:
+            if (
+                device.device_name in appList.keys()
+                or (
+                    "serialNumber" in device.hardware_info
+                    and device.hardware_info["serialNumber"] in appList.keys()
+                )
+                or (
+                    "customSerialNumber" in device.hardware_info
+                    and device.hardware_info["customSerialNumber"] in appList.keys()
+                )
+            ):
+                package_names = None
+                if device.device_name in appList.keys():
+                    package_names = appList[device.device_name]
+                elif (
+                    "serialNumber" in device.hardware_info
+                    and device.hardware_info["serialNumber"] in appList.keys()
+                ):
+                    package_names = appList[device.hardware_info["serialNumber"]]
+                elif (
+                    "customSerialNumber" in device.hardware_info
+                    and device.hardware_info["customSerialNumber"] in appList.keys()
+                ):
+                    package_names = appList[device.hardware_info["customSerialNumber"]]
+                package_names = package_names.split(",")
+                if package_names:
+                    for package_name in package_names:
+                        if package_name.strip():
+                            t = wxThread.GUIThread(
+                                Globals.frame,
+                                apiCalls.setAppState,
+                                args=(
+                                    device.id,
+                                    package_name.strip(),
+                                    state,
+                                ),
+                                name="setAllAppsState",
+                            )
+                            threads.append(t)
+                            t.start()
+                        limitActiveThreads(threads)
+        t = wxThread.GUIThread(
+            Globals.frame,
+            wxThread.waitTillThreadsFinish,
+            args=(tuple(threads), state, -1, 4),
+            name="waitTillThreadsFinish%s" % state,
+        )
+        t.start()
+
+
+def getDevicesFromGrid(deviceIdentifers=None, maxAttempt=Globals.MAX_RETRY):
+    if not deviceIdentifers:
+        deviceIdentifers = Globals.frame.gridPanel.getDeviceIdentifersFromGrid()
+    devices = []
+    for entry in deviceIdentifers:
+        api_instance = esperclient.DeviceApi(
+            esperclient.ApiClient(Globals.configuration)
+        )
+        api_response = None
+        identifier = None
+        for attempt in range(maxAttempt):
+            try:
+                if entry[0]:
+                    identifier = entry[0]
+                    api_response = api_instance.get_all_devices(
+                        Globals.enterprise_id,
+                        name=identifier,
+                        limit=Globals.limit,
+                        offset=Globals.offset,
+                    )
+                elif entry[1]:
+                    identifier = entry[1]
+                    api_response = api_instance.get_all_devices(
+                        Globals.enterprise_id,
+                        serial=identifier,
+                        limit=Globals.limit,
+                        offset=Globals.offset,
+                    )
+                elif entry[2]:
+                    identifier = entry[2]
+                    api_response = api_instance.get_all_devices(
+                        Globals.enterprise_id,
+                        search=identifier,
+                        limit=Globals.limit,
+                        offset=Globals.offset,
+                    )
+                ApiToolLog().LogApiRequestOccurrence(
+                    "getAllDevices",
+                    api_instance.get_all_devices,
+                    Globals.PRINT_API_LOGS,
+                )
+                break
+            except Exception as e:
+                if attempt == maxAttempt - 1:
+                    ApiToolLog().LogError(e)
+                    raise e
+                time.sleep(Globals.RETRY_SLEEP)
+        if api_response:
+            devices += api_response.results
+            api_response = None
+        else:
+            postEventToFrame(
+                eventUtil.myEVT_LOG,
+                "---> ERROR: Failed to find device with identifer: %s" % identifier,
+            )
+    return devices
+
+
+def relocateDeviceToNewGroup(frame, maxAttempt=Globals.MAX_RETRY):
+    devices = getDevicesFromGrid()
+    newGroupList = frame.gridPanel.getDeviceGroupFromGrid()
+
+    splitResults = splitListIntoChunks(devices)
+    threads = []
+
+    for chunk in splitResults:
+        t = wxThread.GUIThread(
+            frame,
+            processDeviceGroupMove,
+            args=(chunk, newGroupList),
+            name="processDeviceGroupMove",
+        )
+        threads.append(t)
+        t.start()
+        limitActiveThreads(threads)
+
+    t = wxThread.GUIThread(
+        frame,
+        wxThread.waitTillThreadsFinish,
+        args=(tuple(threads), GridActions.MOVE_GROUP.value, -1, 5),
+        name="waitTillThreadsFinish",
+    )
+    t.start()
+
+
+def processDeviceGroupMove(deviceChunk, groupList):
+    groupId = None
+    results = {}
+    for device in deviceChunk:
+        groupName = None
+        if device.device_name in groupList.keys():
+            groupName = groupList[device.device_name]
+        elif (
+            "serialNumber" in device.hardware_info
+            and device.hardware_info["serialNumber"] in groupList.keys()
+        ):
+            groupName = groupList[device.hardware_info["serialNumber"]]
+        elif (
+            "customSerialNumber" in device.hardware_info
+            and device.hardware_info["customSerialNumber"] in groupList.keys()
+        ):
+            groupName = groupList[device.hardware_info["customSerialNumber"]]
+        if groupName:
+            if len(groupName) == 36 and "-" in groupName:
+                resp = apiCalls.moveGroup(groupName, device.id)
+                respText = resp.text if hasattr(resp, "text") else str(resp)
+                results[device.device_name] = {
+                    "Device Name": device.device_name,
+                    "Device Id": device.id,
+                    "Status Code": resp.status_code
+                    if hasattr(resp, "status_code")
+                    else None,
+                    "Response": resp.json() if hasattr(resp, "json") else respText,
+                }
+            else:
+                groups = apiCalls.getAllGroups(name=groupName).results
+                if groups:
+                    groupId = groups[0].id
+                    resp = apiCalls.moveGroup(groupId, device.id)
+                    respText = resp.text if hasattr(resp, "text") else str(resp)
+                    results[device.device_name] = {
+                        "Device Name": device.device_name,
+                        "Device Id": device.id,
+                        "Status Code": resp.status_code
+                        if hasattr(resp, "status_code")
+                        else None,
+                        "Response": resp.json() if hasattr(resp, "json") else respText,
+                    }
+                else:
+                    results[device.device_name] = {
+                        "Device Name": device.device_name,
+                        "Device Id": device.id,
+                        "Error": "Invalid Group Name given, no matches found, '%s'"
+                        % groupName,
+                    }
+        else:
+            results[device.device_name] = {
+                "Device Name": device.device_name,
+                "Device Id": device.id,
+                "Error": "Invalid Group Name given, '%s'" % groupName,
+            }
+
+    if not results:
+        results["error"] = {"Error": "Failed to find devices to move, check endpoint."}
+    return results
+
+
+def installLatestApp(frame):
+    deviceIdentifers = Globals.frame.gridPanel.getDeviceIdentifersFromGrid()
+    gridAppList, _ = Globals.frame.gridPanel.getAppsFromGrid()
+    devices = getDevicesFromGrid(deviceIdentifers=deviceIdentifers)
+
+    splitResults = splitListIntoChunks(devices)
+    threads = []
+
+    for chunk in splitResults:
+        t = wxThread.GUIThread(
+            frame,
+            processInstallLatestAppChunk,
+            args=(chunk, gridAppList),
+            name="processInstallLatestAppChunk",
+        )
+        threads.append(t)
+        t.start()
+        limitActiveThreads(threads)
+
+    t = wxThread.GUIThread(
+        frame,
+        wxThread.waitTillThreadsFinish,
+        args=(tuple(threads), GridActions.INSTALL_LATEST_APP.value, -1, 5),
+        name="waitTillThreadsFinish",
+    )
+    t.start()
+
+
+def processInstallLatestAppChunk(devices, appList):
+    status = []
+    for device in devices:
+        for app in appList:
+            status.append(apiCalls.installAppOnDevices(app, devices=[device.id]))
+    return status
+
+
+def uninstallApp(frame):
+    deviceIdentifers = Globals.frame.gridPanel.getDeviceIdentifersFromGrid()
+    gridAppList, _ = Globals.frame.gridPanel.getAppsFromGrid()
+    devices = getDevicesFromGrid(deviceIdentifers=deviceIdentifers)
+
+    splitResults = splitListIntoChunks(devices)
+    threads = []
+
+    for chunk in splitResults:
+        t = wxThread.GUIThread(
+            frame,
+            processUninstallAppChunk,
+            args=(chunk, gridAppList),
+            name="processUninstallAppChunk",
+        )
+        threads.append(t)
+        t.start()
+        limitActiveThreads(threads)
+
+    t = wxThread.GUIThread(
+        frame,
+        wxThread.waitTillThreadsFinish,
+        args=(tuple(threads), GridActions.UNINSTALL_LISTED_APP.value, -1, 5),
+        name="waitTillThreadsFinish",
+    )
+    t.start()
+
+
+def processUninstallAppChunk(devices, appList):
+    status = []
+    for device in devices:
+        for app in appList:
+            status.append(apiCalls.uninstallAppOnDevice(app, device=[device.id]))
+    return status
